BT40 How To Report Bugs & BT41 Using Clam() to Limit Values
BT40 Learnt of a bug with the aiming function. Changed headers from including other classes to forward declarations.

BT41 Clamped the max elevation of the barrel to its min and max values as well and also clamped the RelativeSpeed to be between -1 and 1. Set the elevation to be changed depending on the pitch of our DeltaRotation.